edit api

+++ routes/api.php
@@ -159,7 +159,7 @@ Route::post('users', function(Request $request) {
 
 Route::get('user/{id}/{api}', function($id,$api) {
     $viewer = User::where('api_token',$api)->get();
-   /* if (Notfication::where('user_id', '=', $id)->where('sender', '=',$viewer[0]->id)->where('itemType', '=', 6)->exists()) {
+    if (Notfication::where('user_id', '=', $id)->where('sender', '=',$viewer[0]->id)->where('itemType', '=', 6)->exists()) {
         Notification::where('user_id', '=', $id)->where('sender', '=',$viewer[0]->id)->where('itemType', '=', 6)->first()->delete();
         Notification::create([
             'user_id' => $id,
@@ -172,16 +172,16 @@ Route::get('user/{id}/{api}', function($id,$api) {
         ]);
         // user found
     }else{
-            Notification::create([
-                    'user_id' => $id,
-                    'item_id' => 0,
-                    'item' => 'เข้ามาดูโปรไฟล์ของคุณ',
-                    'itemType' => 6,
-                    'notificationStatus' => 1,
-                    'notificationTime' => time(),
-                    'sender' => $viewer[0]->id,
-                ]);;
-    }*/
+    Notification::create([
+            'user_id' => $id,
+            'item_id' => 0,
+            'item' => 'เข้ามาดูโปรไฟล์ของคุณ',
+            'itemType' => 6,
+            'notificationStatus' => 1,
+            'notificationTime' => time(),
+            'sender' => $viewer[0]->id,
+        ]);
+    }
 
     $user = User::find($id);
     return $user;
